added more possible types for arguments

+++ src/WFfiltering_adaptive_precomputation.jl
@@ -74,7 +74,7 @@ end
 # end
 
 
-function precompute_next_terms!(last_sm_max, new_sm_max, log_ν_dict::Dict{Tuple{Int64,Int64},Float64}, log_Cmmi_dict::Dict{Tuple{Int64,Int64},Float64}, log_binomial_coeff_dict::Dict{Tuple{Int64,Int64},Float64}, sα, Δt)
+function precompute_next_terms!(last_sm_max::Integer, new_sm_max::Integer, log_ν_dict::Dict{Tuple{Int64,Int64},Float64}, log_Cmmi_dict::Dict{Tuple{Int64,Int64},Float64}, log_binomial_coeff_dict::Dict{Tuple{Int64,Int64},Float64}, sα, Δt)
     if last_sm_max == 0
         log_binomial_coeff_dict[(0,0)] = 0
     end
